Improved DenseTranspose class to handle lists as parameter
DenseTranspose now expects a list of 1 or more layers that are then concatenated. This is especially useful for epn rl, since the middle layer consists of multiple layers that can be easily mirrored in the decoder now by passing them to DenseTranspose.
Note: DenseTranspose is only used when weight_sharing is set to 'True'